mcp-server/requirements.txt
# MCP Server Dependencies

# FastAPI and related
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# GraphQL support
strawberry-graphql[fastapi]==0.215.1

# Database drivers
snowflake-connector-python==3.6.0
duckdb==0.9.2
sqlalchemy==2.0.23

# Data orchestration
dagster==1.5.13
dagster-webserver==1.5.13
dbt-core==1.7.4

# DataHub integration
acryl-datahub==0.12.1.5

# LLM and AI
langchain==0.0.350
langchain-openai==0.0.2
openai==1.6.1

# Utilities
pydantic==2.5.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.26.0

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0
